pyproject.toml
[project]
name = "nf-variable-generator"
version = "1.0.0"
license = "LGPL-2.1-or-later"
requires-python = ">=3.12"
dependencies = [
    "requests",
    "python-dotenv"
]

[project.scripts]
generate = "main:main"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["./"]